pyproject.toml
[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"

[project]
name = "historical-sources-search"
version = "0.1.0"
description = "Aggregate search for historical primary source documents"
readme = "README.md"
authors = [
    { name = "Ryan Echols", email = "ryan@shadylakemedia.com" },
]
requires-python = ">=3.13"
dependencies = [
    "fastapi~=0.116.1",
    "httpx~=0.28.1",
    "playwright~=1.54.0",
    "pydantic-settings>=2.10.1,<3",
    "uvicorn[standard]~=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = [
    "ALL",
    # rules to include even if a parent is listed as ignored:
    "W605", "D419", "G010", "G101", "G2", "RET502", "RET503",
]
extend-ignore = [
    # explicit conflicts with auto-formatter:
    "W191", "E111", "E114", "E117", "D206", "D300", "Q000", "Q001", "Q002", "Q003", "COM812", "COM819", "E501",
    # whitespace and formatting taken care of by pre-commit:
    "W",
    # comments are always fine:
    "TD", "FIX", "ERA",
    # don't care:
    "C90", "D", "DOC", "ANN002", "ANN003", "ANN401", "S104", "S113", "S311", "FBT",
    "B904", "B905", "CPY", "C408", "EM", "G", "RSE", "RET", "TC", "PTH123", "PLR0133",
    "PLR09", "PLR1711", "PLR2004", "TRY003", "TRY301",
    # actually makes code harder to read:
    "UP015", "PT003", "SIM105", "SIM108",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # allow asserts in tests
    "INP001", # don't require `__init__.py` in `tests` dir
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true
suppress-dummy-args = true

[tool.mypy]
mypy_path = "src"
files = ["."]
check_untyped_defs = true

[tool.poe.tasks]
lint-ruff = "uvx ruff check"
lint-mypy = "mypy"
lint = ["lint-ruff", "lint-mypy"]
test = "pytest"
